Create task to run each Friday & generate schedule (#15)
Use node-cron to run a task each Friday in production environments. Within the task,
generate a schedule, add the schedule's events to Google Calendar, and send an SMS to each
chef considered.

Additionally improve test coverage, fix the computation of a chef's availability score,
and add a method to the Google Calendar service for adding a write-access user to a
calendar.