Fix webpack build (#18)
The removal of the express app saw the main entry file change from src/server.ts to src/index.ts, but the webpack config was not updated accordingly. Adjust the config to succeed at building the application, and add the build script to Travis so similar issue will be caught by CI in the future.